Added Customer Dashboard Page / DB Inventory Access File
This commit includes a customerDash.php file which currently looks like
the order page for the customer. A database access file for the
inventory has also been created. A placeholder png was added to the
project with this commit for use until we get further direction on how
images will work.