Order Object Class Migration
moved add to order and create order logic into Order class so we stay
object oriented and minimize duplicate code. The total field in the cart
works...mostly...when a user closes the dashboard browser tab, reopens,
and signs back in the total given is the price of the first item in the
order they were working on. Not sure why yet. May need to implement a
some logic to prevent users from closing and reopening tabs without
fully signing back in with a new session id.